Update CompanyPanel.jsx
[Refactor] Clear localstorage if the link is /directory, such that it will fetch all companies profile and display.

Allow search query in the url to process semantic results and display it accordingly. This will also allow another user from another device to reach the same search results even if no search results was generated before, because it would trigger based on the search query in the url format.